🎨 extracted timescale_curve_backend.py
